feat:unified dashboard Layout
