+++ frontend/src/pages/DashboardLayout.css
@@ -1,3 +1,77 @@
+body {
+  font-family:
+    'Rethink Sans',
+    -apple-system,
+    Roboto,
+    Helvetica,
+    sans-serif;
+}
+
+.dashboard {
+  min-height: 100vh;
+  display: flex;
+  width: calc(100vw - 372px);
+  margin-left: 372px;
+}
+
+@media (max-width: 1024px) {
+  .dashboard {
+    width: 100vw;
+    margin-left: 0;
+    justify-content: center;
+  }
+}
+
+.dashboard-wrapper {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: calc(100vw - 735px);
+  height: 100%;
+}
+
+.dashboard-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 10px;
+  padding: 0 24px;
+  padding-top: 20px;
+  margin: 100px 0 50px 0;
+  height: 100%;
+}
+
+.dashboard-content {
+  width: 100%;
+  gap: 24px;
+  border-radius: 20px;
+  padding: 1rem 0;
+  color: var(--second-primary);
+  text-align: center;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
+
+.dashboard-title {
+  font-size: 32px;
+  font-weight: 600;
+  color: var(--second-primary);
+  text-align: center;
+}
+
+.top-cards-dashboard {
+  display: flex;
+  justify-content: space-between;
+  gap: 8px;
+  padding: 0px;
+  margin: 0px;
+  width: 642px;
+  height: 101px;
+}
+
 .dashboard-error {
   color: var(--error-color);
 }
@@ -108,6 +182,13 @@
 }
 
 @media (max-width: 768px) {
+  .dashboard-content {
+    width: 100%;
+    max-width: 100%;
+    padding: 0.5rem 0;
+    gap: 16px;
+  }
+
   .top-cards-dashboard {
     width: 500px;
     flex-direction: row;
@@ -190,6 +271,9 @@
 }
 
 @media (max-width: 550px) {
+  .dashboard-container {
+    width: 550px;
+  }
 
   .dashboard-title {
     font-size: 18px;
@@ -197,6 +281,12 @@
     line-height: 21.79px;
   }
 
+  .dashboard-content {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    align-items: center;
+  }
 
   .dashboard-info-container {
     display: grid;
