+++ frontend/src/pages/add-deposit/AddDeposit.jsx
@@ -2,16 +2,12 @@ import React, { useState } from 'react';
 import { ReactComponent as HealthIcon } from 'assets/icons/health.svg';
 import { ReactComponent as EthIcon } from 'assets/icons/ethereum.svg';
 import { useAddDeposit } from 'hooks/useAddDeposit';
+import DashboardLayout from '../DashboardLayout';
 import './addDeposit.css';
 import Card from 'components/ui/card/Card';
 import TokenSelector from 'components/ui/token-selector/TokenSelector';
 import { NUMBER_REGEX } from 'utils/regex';
 import { Button } from 'components/ui/custom-button/Button';
-import Sidebar from 'components/layout/sidebar/Sidebar';
-import clockIcon from 'assets/icons/clock.svg';
-import computerIcon from 'assets/icons/computer-icon.svg';
-import depositIcon from 'assets/icons/deposit.svg';
-import withdrawIcon from 'assets/icons/withdraw.svg';
 import useDashboardData from 'hooks/useDashboardData';
 
 export const AddDeposit = () => {
@@ -50,67 +46,54 @@ export const AddDeposit = () => {
     );
   };
 
-  const dashboardItems = [
-    { id: 'dashboard', name: 'Dashboard', link: '/dashboard', icon: computerIcon },
-    { id: 'position_history', name: 'Position History', link: '/dashboard/position-history', icon: clockIcon },
-    { id: 'deposit', name: 'Add Deposit', link: '/dashboard/deposit', icon: depositIcon },
-    { id: 'withdraw', name: 'Withdraw All', link: '/dashboard/withdraw', icon: withdrawIcon },
-  ];
-
   return (
-    <div className="deposit">
-      <Sidebar items={dashboardItems} />
-      <div className="deposit-wrapper">
-        <div className="deposit-container">
-          <h1 className="deposit-title">zkLend Deposit</h1>
-          <div className="main-container-deposit">
-            <div className="top-cards-deposit">
-              <Card
-                label="Health Factor"
-                value={dashboardData?.health_ratio}
-                icon={<HealthIcon className="icon" />}
-              />
-              <Card
-                label="Borrow Balance"
-                value={formatNumber(dashboardData?.borrowed, true)}
-                icon={<EthIcon className="icon" />}
-              />
-            </div>
-          </div>
-          <h1 className="deposit-title2">Please make a deposit</h1>
-          <TokenSelector
-            selectedToken={selectedToken}
-            setSelectedToken={setSelectedToken}
-            className="deposit-token-selector"
+    <DashboardLayout title="Add Deposit">
+      <div className="main-container-deposit">
+        <div className="top-cards-dashboard">
+          <Card
+            label="Health Factor"
+            value={dashboardData?.health_ratio}
+            icon={<HealthIcon className="icon" />}
           />
-          <div className="amount-input-deposit" aria-labelledby="amount-input-label">
-            <input
-              type="text"
-              id="amount-field"
-              value={amount}
-              onChange={handleAmountChange}
-              pattern="^\d*\.?\d*$"
-              className="amount-field-deposit"
-              aria-describedby="currency-symbol"
-              placeholder="0.00"
-              disabled={isLoading}
-            />
-            <span id="currency-symbol" className="currency-deposit">
-              {selectedToken}
-            </span>
-          </div>
-
-          <Button
-            size="lg"
-            className="deposit-btn"
-            variant="primary"
-            onClick={handleDeposit}
-            disabled={isLoading || amount === '0'}
-          >
-            {isLoading ? 'Processing...' : 'Deposit'}
-          </Button>
+          <Card
+            label="Borrow Balance"
+            value={formatNumber(dashboardData?.borrowed, true)}
+            icon={<EthIcon className="icon" />}
+          />
         </div>
       </div>
-    </div>
+      <h1 className="deposit-title2">Please make a deposit</h1>
+      <TokenSelector
+        selectedToken={selectedToken}
+        setSelectedToken={setSelectedToken}
+        className="deposit-token-selector"
+      />
+      <div className="amount-input-deposit" aria-labelledby="amount-input-label">
+        <input
+          type="text"
+          id="amount-field"
+          value={amount}
+          onChange={handleAmountChange}
+          pattern="^\d*\.?\d*$"
+          className="amount-field-deposit"
+          aria-describedby="currency-symbol"
+          placeholder="0.00"
+          disabled={isLoading}
+        />
+        <span id="currency-symbol" className="currency-deposit">
+          {selectedToken}
+        </span>
+      </div>
+
+      <Button
+        size="lg"
+        className="redeem-btn"
+        variant="primary"
+        onClick={handleDeposit}
+        disabled={isLoading || amount === '0'}
+      >
+        {isLoading ? 'Processing...' : 'Deposit'}
+      </Button>
+    </DashboardLayout>
   );
 };
